fixes #285 (second part)
