+++ dbptk-bindings/pom.xml
@@ -9,7 +9,7 @@
     <parent>
         <groupId>com.databasepreservation</groupId>
         <artifactId>dbptk</artifactId>
-        <version>2.0.0-beta6.1</version>
+        <version>2.0.0-beta6.2</version>
         <relativePath>..</relativePath>
     </parent>
     <packaging>pom</packaging>
