+++ dbptk-model/pom.xml
@@ -5,12 +5,12 @@
     <name>dbptk-model</name>
     <groupId>com.databasepreservation</groupId>
     <artifactId>dbptk-model</artifactId>
-    <version>2.0.0-beta6.1</version>
+    <version>2.0.0-beta6.2</version>
     <packaging>jar</packaging>
     <parent>
         <groupId>com.databasepreservation</groupId>
         <artifactId>dbptk</artifactId>
-        <version>2.0.0-beta6.1</version>
+        <version>2.0.0-beta6.2</version>
         <relativePath>..</relativePath>
     </parent>
     <dependencies>
