+++ dbptk-modules/dbptk-module-solr/pom.xml
@@ -6,11 +6,11 @@
     <name>dbptk-module-solr</name>
     <groupId>com.databasepreservation</groupId>
     <artifactId>dbptk-module-solr</artifactId>
-    <version>2.0.0-beta6.1</version>
+    <version>2.0.0-beta6.2</version>
     <parent>
         <groupId>com.databasepreservation</groupId>
         <artifactId>dbptk-modules</artifactId>
-        <version>2.0.0-beta6.1</version>
+        <version>2.0.0-beta6.2</version>
         <relativePath>..</relativePath>
     </parent>
