+++ dbptk-modules/pom.xml
@@ -5,11 +5,11 @@
     <name>dbptk-modules</name>
     <groupId>com.databasepreservation</groupId>
     <artifactId>dbptk-modules</artifactId>
-    <version>2.0.0-beta6.1</version>
+    <version>2.0.0-beta6.2</version>
     <parent>
         <groupId>com.databasepreservation</groupId>
         <artifactId>dbptk</artifactId>
-        <version>2.0.0-beta6.1</version>
+        <version>2.0.0-beta6.2</version>
         <relativePath>..</relativePath>
     </parent>
     <packaging>pom</packaging>
